+++ Changes.md
@@ -1,3 +1,4 @@
+* #850 - Add support for delaying the display of performance scores
 * #821 - Reformat the finalist team schedule to be similar to the grid used for scheduling
 * #737 - Validate that all goals of a subjective category have the same rubric range titles
 * #739 - Validate that all goals of a subjective category have the same rubric range min and max values